added gernerationLog class and merged features from old class crossover is broken(!)

+++ src/GenerationHandler.java
@@ -5,6 +5,7 @@ import java.io.PrintWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
@@ -16,12 +17,11 @@ public class GenerationHandler {
     private ArrayList<Kette> individuals = new ArrayList<>();
     private ImageCreator imageCreator = new ImageCreator();
     private RandomCollection<Kette> randomCollection = new RandomCollection<>();
+    private int generation;
     private int maxGenerations = 0;
     private int generationSize = 0;
     private int newBloodAmount = 0;
-    private Log Log;
-    private Kette bestIndividual = new Kette("");
-
+    private generationLog log = new generationLog();
 
     public GenerationHandler(String sequence) throws IOException {
         this.sequence = sequence;
@@ -32,7 +32,7 @@ public class GenerationHandler {
             //noinspection ResultOfMethodCallIgnored
             new File(folder).mkdirs();
         }
-        else{
+        else if(new File("/ga/!Log.txt").exists()){
             Path fileToDeletePath = Paths.get("/ga/!Log.txt");
             Files.delete(fileToDeletePath);
         }
@@ -40,43 +40,54 @@ public class GenerationHandler {
 
 
     void initializeGeneration(int generationSize) {
+        this.generationSize = generationSize;
         for (int i = 0; i < generationSize; i++) {
             individuals.add(new Kette(sequence));
             individuals.get(i).generateByRng();
         }
+        log.saveGeneration(individuals);
+        printLogTxt();
     }
 
-    void evolve(int maxGenerations, int generationSize, int newBloodAmount){
+    void evolve(int maxGenerations, int newBloodAmount){
         this.maxGenerations = maxGenerations;
-        this.generationSize = generationSize;
         this.newBloodAmount = newBloodAmount;
 
 
-        for (int generation = 0; generation < maxGenerations; generation++){
-            //fitnessBiasedSelection(generationSize/2);
-            tournamentSelection(10,50);
-            //individuals.sort((Kette ketteA, Kette ketteB) -> Double.compare(ketteB.calcFitness(),ketteA.calcFitness()));
+        for (this.generation = 1; this.generation < this.maxGenerations; this.generation++){
 
-            printLogTxt(generation);
-            if (generation != maxGenerations -1) { // if not the last generation
+            // selection Process
+            fitnessBiasedSelection(generationSize);
+            //tournamentSelection(10,50);
+            //individuals.sort((Kette ketteA, Kette ketteB) -> Double.compare(ketteB.calcFitness(),ketteA.calcFitness()));
+            log.saveGeneration(individuals);
+            printLogTxt();
+
+            // mutation, crossover etc.
+            if (this.generation != this.maxGenerations -1) { // if not the last generation
                 //makeSomeBabys();
-                makeSomeMutants(generation);
-                //makeSomeNewBlood(generation);
+                //crossover(0.25); //Broken!
+                mutation(1);
+                                //makeSomeNewBlood(generation);
+                shitCleanUp();
             }else{ //if its the last generation
                 //individuals.subList(5, individuals.size()).clear(); // kill all but the x best
             }
-            printLogTxt();
-            createImageOfTheBestIn(generation);
+
+            createImageOfTheBestIn();
         }
+        log.saveGeneration(individuals);
+        printLogTxt();
     }
+
 
     // https://www.youtube.com/watch?v=9JzFcGdpT8E
     private int fitnessBiasedSelection() {
         Random rng = new Random();
-        double rand = Log.getCurrentFitness() * rng.nextDouble();
+        double rand = log.getSumOfFintessIn(this.generation) * rng.nextDouble();
         double partialSum = 0;
         for (int x = generationSize - 1; x >= 0; x--) {
-            partialSum += individuals.get(x).calcFitness();
+            partialSum += individuals.get(x).getFitness();
             if (partialSum >= rand) {
                 return x;
             }
@@ -92,14 +103,16 @@ public class GenerationHandler {
     }
 
     private void shitCleanUp(){
-        for(int j = 0; j < generationSize; j++){
-            individuals.remove(j);
+        int counter = 0;
+        while(individuals.size() != generationSize){
+            individuals.remove(counter);
+        counter++;
         }
     }
 
-    private void crossover(){ //Todo refactor!!!
+    private void crossover(double rate){ //Todo refactor!!!
         //create 2 offspring's
-        for (int i = 0; i < generationSize / 8; i++) {
+        for (int i = 0; i < ((generationSize * rate/2)); i++) {
             ArrayList<Integer> chromosomeA = ChromosomeHandler.extractChromosome(individuals.get(fitnessBiasedSelection()).getPhenotype());
             ArrayList<Integer> chromosomeB = ChromosomeHandler.extractChromosome(individuals.get(fitnessBiasedSelection()).getPhenotype());
 
@@ -142,7 +155,7 @@ public class GenerationHandler {
     }
 
     private void generateRandomCollection() {
-        double overallFitness = calcOverallFitness();
+        double overallFitness = log.getSumOfFintessIn(this.generation -1);
         for (Kette individual : individuals){
             double weight = (individual.calcFitness()/overallFitness);
             weight = weight*100;
@@ -163,10 +176,10 @@ public class GenerationHandler {
     }
 
     //Todo: make altering the mutation rate somewhat convenient
-    private void makeSomeMutants(int generation){
+    private void mutation(double rate){
         int initialPop = individuals.size();
         // fill the generationSize while leaving space for newBlood also no need to do that in the last gen
-        while (individuals.size() < generationSize - newBloodAmount){
+        while (individuals.size() < (generationSize * rate - newBloodAmount) + generationSize){
             int randomNum = ThreadLocalRandom.current().nextInt(0, initialPop);
             ArrayList<Integer> chromosomeMutant = ChromosomeHandler.extractChromosome(individuals.get(randomNum).getPhenotype());
             ArrayList<Integer> mutant = ChromosomeHandler.mutateChromosome(chromosomeMutant, 0.1);
@@ -174,21 +187,13 @@ public class GenerationHandler {
         }
     }
 
-    private void makeSomeNewBlood(int generation){
+    private void makeSomeNewBlood(){
         while (individuals.size() < generationSize){
             individuals.add(new Kette(sequence));
             individuals.get(individuals.size()-1).generateByRng();
         }
     }
 
-    private double calcOverallFitness(){
-        double avr = 0;
-        for (Kette kette : individuals){
-            avr += kette.calcFitness();
-        }
-        //avr = avr/individuals.size();
-        return avr;
-    }
 
     void drawResult(int top) { // top defines the best x you want the image of
         individuals.sort((Kette ketteA, Kette ketteB) -> Double.compare(ketteB.calcFitness(),ketteA.calcFitness()));
@@ -199,25 +204,19 @@ public class GenerationHandler {
         }
     }
 
-    private void printLogTxt(int generation){
-        individuals.sort((Kette ketteA, Kette ketteB) -> Double.compare(ketteB.calcFitness(),ketteA.calcFitness()));
-        if(bestIndividual.calcFitness() < individuals.get(0).calcFitness())
-            bestIndividual = individuals.get(0);
-        //create folder
-        String folder = "/ga";
-        if (!new File(folder).exists()) {
-            //noinspection ResultOfMethodCallIgnored
-            new File(folder).mkdirs();
-        }
-        try (PrintWriter out = new PrintWriter(new FileWriter(new File("/ga" + File.separator +"Log.txt"),true))) {
-            out.print((Integer.toString(generation) + "," + calcOverallFitness() / generationSize) + "," +
-                    individuals.get(0).calcFitness() + "," + bestIndividual.calcFitness() + "," +
-                    bestIndividual.calcMinEnergy() + "," + bestIndividual.calcOverlap());
+    private void printLogTxt(){
+        try (PrintWriter out = new PrintWriter(new FileWriter(new File("/ga" + File.separator +"!Log.txt"),true))) {
+            out.print((Integer.toString(this.generation) + "," + log.getAverageFitnessIn(this.generation)) + "," +
+                    log.getGenerationsBestFitnessIn(this.generation) + "," + log.bestIndividual.calcFitness() + "," +
+                    log.bestIndividual.calcMinEnergy() + "," + log.bestIndividual.calcOverlap());
 
             out.print("\n");
         }catch (java.io.IOException e){
             System.out.println("Log file not found");
         }
+        System.out.print(MessageFormat.format("Generation: {0} \t Average: {1} \t Best: {2} \n", Integer.toString(this.generation),
+                Double.toString(log.getAverageFitnessIn(this.generation)), Double.toString(log.getGenerationsBestFitnessIn(this.generation))));
+
     }
 
     private static int getRandomIntInRange(int min, int max) {
@@ -231,11 +230,10 @@ public class GenerationHandler {
     }
 
 
-    private void createImageOfTheBestIn(int generation){
-
+    private void createImageOfTheBestIn(){
         individuals.sort((Kette ketteA, Kette ketteB) -> Double.compare(ketteB.calcFitness(),ketteA.calcFitness()));
-        imageCreator.createImage(individuals.get(0).getPhenotype(), "Generation_" + Integer.toString(generation)+ ".png");
-        System.out.println();
+        imageCreator.createImage(individuals.get(0).getPhenotype(), "Generation_" + Integer.toString(this.generation)+ ".png");
+        //System.out.println();
         //individuals.get(0).printValues();
     }
